+++ src/libs/int/int.h
@@ -1,5 +1,8 @@
 #pragma once
 
-char *uint_to_string(unsigned int self);
-char *uint_to_subscript(unsigned int self);
-char *uint_to_superscript(unsigned int self);
+// bool uint_max(uint64_t x, uint64_t y);
+// bool uint_min(uint64_t x, uint64_t y);
+
+char *uint_to_string(uint64_t self);
+char *uint_to_subscript(uint64_t self);
+char *uint_to_superscript(uint64_t self);
